+++ frontend/solstar/src/components/funding/regist/ArtistSearchModal.css
