+++ frontend/solstar/src/components/funding/regist/FundingRegistContent.css
