+++ frontend/solstar/src/components/funding/regist/FundingRegistContent.jsx
