+++ frontend/solstar/src/components/funding/regist/FundingRegistInfo.css
