+++ frontend/solstar/src/components/funding/search/FundingSearchResult.jsx
@@ -1,0 +1,3 @@
+function FundingSearchResult() {}
+
+export default FundingSearchResult;
